refactor(#792): split param_tests into base + strict/builders/coercion suite (sweep)
param_tests.rs 1102 -> 588, + param_tests_more.rs 520 (both under cap; off debt).
The strict-mode / builder-fn / coercion tests move to the sibling; test_schema
becomes pub(super), imported via super::tests. block-core tests pass, zero warnings.
